python/discord.py/subclassed-bot/requirements.txt
discord.py>=1.7.0
